Chore(rename): Rename EC2-Terminate experiments to EC2-Stop
