+++ chaoslib/litmus/ec2-stop-by-id/lib/ec2-stop-by-id.go
@@ -1,25 +1,22 @@
 package lib
 
 import (
-	"math/rand"
 	"strings"
 	"time"
 
 	clients "github.com/litmuschaos/litmus-go/pkg/clients"
 	awslib "github.com/litmuschaos/litmus-go/pkg/cloud/aws"
 	"github.com/litmuschaos/litmus-go/pkg/events"
-	experimentTypes "github.com/litmuschaos/litmus-go/pkg/kube-aws/ec2-terminate-by-tag/types"
+	experimentTypes "github.com/litmuschaos/litmus-go/pkg/kube-aws/ec2-stop-by-id/types"
 	"github.com/litmuschaos/litmus-go/pkg/log"
-	"github.com/litmuschaos/litmus-go/pkg/math"
 	"github.com/litmuschaos/litmus-go/pkg/probe"
 	"github.com/litmuschaos/litmus-go/pkg/types"
 	"github.com/litmuschaos/litmus-go/pkg/utils/common"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 )
 
-//PrepareEC2TerminateByTag contains the prepration and injection steps for the experiment
-func PrepareEC2TerminateByTag(experimentsDetails *experimentTypes.ExperimentDetails, clients clients.ClientSets, resultDetails *types.ResultDetails, eventsDetails *types.EventDetails, chaosDetails *types.ChaosDetails) error {
+//PrepareEC2StopByID contains the prepration and injection steps for the experiment
+func PrepareEC2StopByID(experimentsDetails *experimentTypes.ExperimentDetails, clients clients.ClientSets, resultDetails *types.ResultDetails, eventsDetails *types.EventDetails, chaosDetails *types.ChaosDetails) error {
 
 	var err error
 	//Waiting for the ramp time before chaos injection
@@ -28,8 +25,11 @@ func PrepareEC2TerminateByTag(experimentsDetails *experimentTypes.ExperimentDeta
 		common.WaitForDuration(experimentsDetails.RampTime)
 	}
 
-	instanceIDList := CalculateInstanceAffPerc(experimentsDetails.InstanceAffectedPerc, experimentsDetails.TargetInstanceIDList)
-	log.Infof("[Chaos]:Number of Instance targeted: %v", len(instanceIDList))
+	//get the instance id or list of instance ids
+	instanceIDList := strings.Split(experimentsDetails.Ec2InstanceID, ",")
+	if len(instanceIDList) == 0 {
+		return errors.Errorf("no instance id found to stop")
+	}
 
 	if strings.ToLower(experimentsDetails.Sequence) == "serial" {
 		if err = InjectChaosInSerialMode(experimentsDetails, instanceIDList, clients, resultDetails, eventsDetails, chaosDetails); err != nil {
@@ -48,7 +48,7 @@ func PrepareEC2TerminateByTag(experimentsDetails *experimentTypes.ExperimentDeta
 	return nil
 }
 
-//InjectChaosInSerialMode will inject the ce2 instance termination in serial mode that is one after other
+//InjectChaosInSerialMode will inject the ec2 instance termination in serial mode that is one after other
 func InjectChaosInSerialMode(experimentsDetails *experimentTypes.ExperimentDetails, instanceIDList []string, clients clients.ClientSets, resultDetails *types.ResultDetails, eventsDetails *types.EventDetails, chaosDetails *types.ChaosDetails) error {
 
 	//ChaosStartTimeStamp contains the start timestamp, when the chaos injection begin
@@ -125,7 +125,7 @@ loop:
 	return nil
 }
 
-// InjectChaosInParallelMode will inject the ce2 instance termination in parallel mode that is all at once
+// InjectChaosInParallelMode will inject the ec2 instance termination in parallel mode that is all at once
 func InjectChaosInParallelMode(experimentsDetails *experimentTypes.ExperimentDetails, instanceIDList []string, clients clients.ClientSets, resultDetails *types.ResultDetails, eventsDetails *types.EventDetails, chaosDetails *types.ChaosDetails) error {
 
 	//ChaosStartTimeStamp contains the start timestamp, when the chaos injection begin
@@ -185,6 +185,7 @@ loop:
 			for _, id := range instanceIDList {
 				//Wait for ec2 instance to get in running state
 				log.Infof("[Wait]: Wait for EC2 instance '%v' to get in running state", id)
+				experimentsDetails.Ec2InstanceID = id
 				if err := awslib.WaitForEC2Up(experimentsDetails.Timeout, experimentsDetails.Delay, experimentsDetails.ManagedNodegroup, experimentsDetails.Region, id); err != nil {
 					return errors.Errorf("unable to start the ec2 instance, err: %v", err)
 				}
@@ -207,66 +208,22 @@ loop:
 	return nil
 }
 
-//CalculateInstanceAffPerc will calculate the target instance ids according to the instance affected percentage provided.
-func CalculateInstanceAffPerc(InstanceAffPerc int, instanceList []string) []string {
+//InstanceStatusCheckByID is used to check the instance status of all the instance under chaos.
+func InstanceStatusCheckByID(experimentsDetails *experimentTypes.ExperimentDetails) error {
 
-	var newIDList []string
-	newInstanceListLength := math.Maximum(1, math.Adjustment(InstanceAffPerc, len(instanceList)))
-	rand.Seed(time.Now().UnixNano())
-
-	// it will generate the random instanceList
-	// it starts from the random index and choose requirement no of instanceID next to that index in a circular way.
-	index := rand.Intn(len(instanceList))
-	for i := 0; i < newInstanceListLength; i++ {
-		newIDList = append(newIDList, instanceList[index])
-		index = (index + 1) % len(instanceList)
+	instanceIDList := strings.Split(experimentsDetails.Ec2InstanceID, ",")
+	if len(instanceIDList) == 0 {
+		return errors.Errorf("no instance id found to stop")
 	}
-	return newIDList
-}
-
-//PostChaosInstanceStatusCheck is used to check the instance status of the instances under chaos post chaos.
-func PostChaosInstanceStatusCheck(targetInstanceIDList []string, region string) error {
-
-	for _, id := range targetInstanceIDList {
-		instanceState, err := awslib.GetEC2InstanceStatus(id, region)
+	log.Infof("[Info]: The instances under chaos(IUC) are: %v", instanceIDList)
+	for _, id := range instanceIDList {
+		instanceState, err := awslib.GetEC2InstanceStatus(id, experimentsDetails.Region)
 		if err != nil {
 			return err
 		}
 		if instanceState != "running" {
-			return errors.Errorf("failed to get the ec2 instance '%v' in running sate, current state: %v", id, instanceState)
+			return errors.Errorf("failed to get the ec2 instance '%v' status as running", id)
 		}
 	}
-	return nil
-}
-
-//SetTargetInstance will select the target instance which are in running state and filtered from the given instance tag
-func SetTargetInstance(experimentsDetails *experimentTypes.ExperimentDetails) error {
-
-	instanceIDList, err := awslib.GetInstanceList(experimentsDetails.InstanceTag, experimentsDetails.Region)
-	if err != nil {
-		return err
-	}
-	if len(instanceIDList) == 0 {
-		return errors.Errorf("no instance found with the given tag %v, in region %v", experimentsDetails.InstanceTag, experimentsDetails.Region)
-	}
-
-	for _, id := range instanceIDList {
-		instanceState, err := awslib.GetEC2InstanceStatus(id, experimentsDetails.Region)
-		if err != nil {
-			return errors.Errorf("fail to get the instance status while selecting the target instances, err: %v", err)
-		}
-		if instanceState == "running" {
-			experimentsDetails.TargetInstanceIDList = append(experimentsDetails.TargetInstanceIDList, id)
-		}
-	}
-
-	if len(experimentsDetails.TargetInstanceIDList) == 0 {
-		return errors.Errorf("fail to get any running instance having instance tag: %v", experimentsDetails.InstanceTag)
-	}
-
-	log.InfoWithValues("[Info]: Targeting the running instances filtered from instance tag", logrus.Fields{
-		"Total number of instances filtered":   len(instanceIDList),
-		"Number of running instances filtered": len(experimentsDetails.TargetInstanceIDList),
-	})
 	return nil
 }
