+++ experiments/kube-aws/ec2-stop-by-id/README.md
@@ -7,8 +7,8 @@
 <th> Documentation Link </th>
 </tr>
 <tr>
- <td> EC2 Terminate </td>
+ <td> EC2 Stop By ID </td>
  <td> This experiment causes termination of an EC2 instance before bringing it back to running state after the specified chaos duration. One or more target instance is provided in the list format in `EC2_INSTANCE_ID` env as comma(,) seperated envs (eg: instance1,instance2)</td>
- <td>  <a href=""> Coming Soon </a> </td>
+ <td>  <a href="https://docs.litmuschaos.io/docs/ec2-terminate-by-id/"> Here </a> </td>
  </tr>
  </table>
