+++ experiments/kube-aws/ec2-stop-by-id/rbac.yaml
@@ -2,18 +2,18 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: ec2-terminate-by-tag-sa
+  name: ec2-stop-by-id-sa
   namespace: default
   labels:
-    name: ec2-terminate-by-tag-sa
+    name: ec2-stop-by-id-sa
     app.kubernetes.io/part-of: litmus
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
-  name: ec2-terminate-by-tag-sa
+  name: ec2-stop-by-id-sa
   labels:
-    name: ec2-terminate-by-tag-sa
+    name: ec2-stop-by-id-sa
     app.kubernetes.io/part-of: litmus
 rules:
 - apiGroups: [""]
@@ -35,15 +35,15 @@ rules:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: ec2-terminate-by-tag-sa
+  name: ec2-stop-by-id-sa
   labels:
-    name: ec2-terminate-by-tag-sa
+    name: ec2-stop-by-id-sa
     app.kubernetes.io/part-of: litmus
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: ec2-terminate-by-tag-sa
+  name: ec2-stop-by-id-sa
 subjects:
 - kind: ServiceAccount
-  name: ec2-terminate-by-tag-sa
+  name: ec2-stop-by-id-sa
   namespace: default
