+++ experiments/kube-aws/ec2-stop-by-id/test/test.yml
@@ -13,7 +13,7 @@ spec:
       labels:
         app: litmus-experiment
     spec:
-      serviceAccountName: ec2-terminate-by-id-sa
+      serviceAccountName: ec2-stop-by-id-sa
       containers:
       - name: gotest
         image: busybox
