+++ experiments/kube-aws/ec2-stop-by-tag/README.md
@@ -9,6 +9,6 @@
 <tr>
  <td> EC2 Terminate By Tag </td>
  <td> This experiment causes termination of an EC2 instance before bringing it back to running state using the instance tag after the specified chaos duration. We can also control the number of target instance using instance affected percentage</td>
- <td>  <a href=""> Coming Soon </a> </td>
+ <td>  <a href="https://docs.litmuschaos.io/docs/ec2-terminate-by-tag/"> Here </a> </td>
  </tr>
  </table>
