+++ experiments/kube-aws/ec2-stop-by-tag/rbac.yaml
@@ -2,18 +2,18 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: ec2-terminate-by-id-sa
+  name: ec2-stop-by-tag-sa
   namespace: default
   labels:
-    name: ec2-terminate-by-id-sa
+    name: ec2-stop-by-tag-sa
     app.kubernetes.io/part-of: litmus
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
-  name: ec2-terminate-by-id-sa
+  name: ec2-stop-by-tag-sa
   labels:
-    name: ec2-terminate-by-id-sa
+    name: ec2-stop-by-tag-sa
     app.kubernetes.io/part-of: litmus
 rules:
 - apiGroups: [""]
@@ -35,15 +35,15 @@ rules:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: ec2-terminate-by-id-sa
+  name: ec2-stop-by-tag-sa
   labels:
-    name: ec2-terminate-by-id-sa
+    name: ec2-stop-by-tag-sa
     app.kubernetes.io/part-of: litmus
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: ec2-terminate-by-id-sa
+  name: ec2-stop-by-tag-sa
 subjects:
 - kind: ServiceAccount
-  name: ec2-terminate-by-id-sa
+  name: ec2-stop-by-tag-sa
   namespace: default
