+++ pkg/kube-aws/ec2-stop-by-id/types/types.go
