+++ pkg/kube-aws/ec2-stop-by-tag/types/types.go
